break up prompts, validators and inputs
