+++ packages/flutterfire_cli/lib/src/commands/config.dart
@@ -486,7 +486,7 @@ class ConfigCommand extends FlutterFireCommand {
       );
     }
 
-    final firebaseConfigurationFileInputs = firebaseConfigurationFileValidation(
+    final firebaseConfigurationFileInputs = dartConfigurationFileValidation(
       flutterAppPath: flutterApp!.package.path,
     );
